[Chore] Break tests to smaller chunks
